+++ module/helpers/chat.mjs
@@ -1,4 +1,7 @@
 export function chatListeners(message, html) {
+//  html.on("click", "button.dmgroll", _onDmgRollClick.c(this));
+  html.on("click", "button.dmgroll", (event) => _onDmgRollClick.call(this, event, message));
+
   html.on("click", ".card-buttons button", _onChatCardAction.bind(this));
   // Add reroll buttons to all dice rolls
   html.find(".roll").each((_i, div) => {
@@ -417,6 +420,57 @@ export function _findCharTargets() {
     chars.push(game.user.character);
   }
   return chars;
+}
+
+// Load a saved damage roll from the message flags
+export async function _onDmgRollClick(event, message) {
+  event.preventDefault();
+  const btn = event.currentTarget;
+  const ns  = btn.dataset.flagns;
+  const key = btn.dataset.flagkey;
+
+  const payload = message.getFlag(ns, key);
+  if (!payload?.formula) return;
+
+  const actor = payload.actorId ? game.actors.get(payload.actorId) : null;
+
+  const damageRoll = await (new Roll(payload.formula, payload.data)).evaluate({ async: true });
+
+  let traumaRollRender = null;
+  let traumaDamage = null;
+  if (payload.traumaFormula) {
+    const traumaRoll = new Roll(payload.traumaFormula);
+    await traumaRoll.roll();
+    traumaRollRender = await traumaRoll.render();
+    if (traumaRoll && traumaRoll.total && traumaRoll.total >= 6 && damageRoll?.total) {
+      traumaDamage = new Roll(`${payload.traumaRating} * ${damageRoll.total}`);
+      await traumaDamage.roll();
+      traumaDamage = await traumaDamage.render();
+    }
+  }
+  const rollMode = game.settings.get("core", "rollMode");
+
+  const damageRollTemplate = "systems/swnr/templates/chat/damage-roll.hbs";
+  const damageRollData = {
+    actor: actor,
+    weapon: payload.weaponId ? game.items.get(payload.weaponId) : null,
+    damageRoll: damageRoll,
+    damage: await damageRoll.render(),
+    damageExplain: payload.damageExplain,
+    traumaRollRender,
+    traumaDamage,
+  };
+  const damageRollContent = await renderTemplate(damageRollTemplate, damageRollData);
+  const chatData = {
+    speaker: ChatMessage.getSpeaker({ actor }),
+    content: damageRollContent,
+    roll: JSON.stringify(damageRoll),
+    rolls: [damageRoll],
+    content: damageRollContent,
+    flavor: payload.flavor,
+  };
+  getDocumentClass("ChatMessage").applyRollMode(chatData, rollMode);
+  getDocumentClass("ChatMessage").create(chatData);
 }
 
 //Taken from WWN (which could have come from OSE)