+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to the Systems Without Number Redux (SWNR) system for Foundr
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.0] 2025-11-04 More XWN support
+
+- Custom currency system configuration supported. Base currency and up to 5 custom currencies. Old debt, balance, and owed fields are deprecated and will be removed in a future version.  They should be migrated to the new system, but the old values are shown in the tweaks section as readonly.
+- Added readied and stowed modifiers to character sheet (under features/tweaks)
+- Fixed capt support department bug
+- Added stress button to character sheet to modify and roll stress, along with tracking breakdowns (page 57 of AWN).
+- Auto show power type on features tab when type is first added
+- Setting to not roll damage/trauma dice on attack roll automatically with button to roll
+- Added GM notes to item description (under description tab). GM notes are hidden by default and can be shown to players by checking the box.
+
 ## [2.2.1] 2025-11-11 CWN Cyberware compendium
 
 - Re-added CWN Cyberware compendium
@@ -22,7 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed issues for ship / mech attack
 - Removed roll formula from item description by default
 - Updated WWN / OSE compendium content
-- Updated NPC import from CSV
+- Updated NPC import from CSV for compendium import
 - Added condition to item description and chat messages. Not yet automated. Behind AWN setting.
 - Shock damage now uses dice string
 - Changed several fields to validate that they are valid dice strings (such as damage, trauma, etc)